requirements.txt
numpy
Pillow
matplotlib
opencv-python
tensorflow
torch
torchvision
opencv-python
pytorch-lightning
albumentations
pandas
matplotlib
segmentation_models_pytorch
ultralytics

